modules/grpc/requirements.txt
grpcio
grpcio-tools
kafka-python==2.0.2